Modified data model and terms to remove ambiguity
This PR tries to remove ambiguity from the data model and terms by doing the following things:

i) makes sure that each JSON structure is clearly identified for what it is (i.e. verifiable credential or verifiable profile). A side benefit of this is that the type parameter can be simplified as it no longer needs to specify the base type.

ii) removed inconsistencies whereas sometimes a VC was called a credential and sometimes a verifiableCredential

iii) made the "id" property consistent in that it always comes first in a data type definition and always uniquely identifies that data type

iv) updated the terms to make it clear that a verifiable credential holds a set of claims, that each claim contains the properties of one subject, and that the subject can be different for each claim. Each subject has a unique subject id and each subject comprises a set of properties.

v) that each credential status scheme is clearly identified in the credential status property, and each scheme is responsible for defining the contents of the status property

vi) for verifiable profiles, added the holder id, and made it clear that the existing id is the profile id (consistent with i) above

vii) for terms of use, changed the type into a unique type ID, and made the rest of the contents be dependent upon this type. Consequently new types of terms of use can be defined with different properties to the current examples.

viii) turned evidence into a set, consistent with the existing text, with each element being identified by its unique scheme ID. The contents of the scheme properties are then determined by the scheme itself

ix) removed revocation and replaced with credential status