Grammar/typo/formatting updates for Section 1
Mostly grammatical, punctuation, and formatting changes. I began capitalizing the first word in a bullet point but don't know if you have a specific style rule for this or not.
As I always say in my reviews, my suggestions are exactly that. Suggestions. You can accept, all, some, or none of my suggestions. I'm not particulartly precious about them. :smile: