Section 3.2 links and grammar and other suggestions
The text describing the information graphs has some "funky" styling that might not be appropriate for the tone of the document? :) Perhaps suggest just using <em> tags instead?
Also, I removed a link from a figure title. Usually terms in headings and in figure and table labels are not linked? At least that seems to be the convention so far.