Verification section updates
to clarify holder vs subject. For the most part I believe this section is ok, however I'm suggesting some changes to tighten up what we are talking about. Specifically, verifying verifiable presentations (which contain verifiable credentials) and in a few instances, clarify that it's the holder making the presentation on behalf of a subject.