Appendix C Subject Holder suggested editorial updates
The sentences before and after the diagram seemed to be saying very much the same thing, so incorporated them into the opening para and deleted the second one.
Some wording changes, and some punctuation.
In "Issuer Authorizes Holder" section, the third para was an almost word for word duplication of the opening para. Deleted the third para.