Better not set §B as non-normative (minor editorial)
At the moment §B declares itself as non-normative, and this is, sort of, "inherited" by subsections unless stated otherwise. Ie, §B.1 and §B.2 may be considered as non-normative, which is not the intention. Better leave §B to normative (the default) and §B.3 explicitly declares itself as non-normative.

Just to avoid misunderstandings...