Use JSON-LD keywords for language/direction.
Aliasing these keywords confuses their use and presents them as acting
as any other term--and implying additional terms may exist within these
objects. The `@` prefixes provide a clearer distinction that an object
shaped in this fashion has a unique and singular purpose of expressing a
string (in `@value`) and related metadata (in `@language` and/or
`@direction`).